+++ tests/test_pluto_metrics.c
@@ -17,8 +17,8 @@ int main(int argc, const char* argv[])
     size_t nfft;
     int res, navg;
     double fs, fsr;
-    err_code = read_param(test_filename, "domain_wf", (void*)(&domain_wf), UINT64);
-    err_code = read_param(test_filename, "type_wf", (void*)(&type_wf), UINT64);
+    err_code = read_param(test_filename, "domain_wf", (void*)(&domain_wf), UINT32);
+    err_code = read_param(test_filename, "type_wf", (void*)(&type_wf), UINT32);
     err_code = read_param(test_filename, "nfft", (void*)(&nfft), UINT64);
     err_code = read_param(test_filename, "res", (void*)(&res), INT32);
     err_code = read_param(test_filename, "navg", (void*)(&navg), INT32);
